add gradle and travis and make android library project
